Removed unneeded dependency (#58)
* Removed unneeded dependency.

This commit removes the unnecessary github.com/katzenpost/core dependency. We were only
using this library for a few helper functions, which are trivially implemented using
the standard library. I also added some tests to confirm these changes.

This also has the side effect of fixing our broken implementation of DisplayPublicKey
for OKPs, since the initial implementation sent the public key bytes to the `ToPEMFile`
function which expected a filename.

I've updated go.mod and go.sum using "go mod tidy" to reflect the new changes.

* Added legacy support for marshaling Ed25519 public keys.

On further review, it appears that support for Ed25519 public keys in
MarshalPKIXPublicKey was added in Go 1.13. I backported the changes so that we
can continue to support older versions of Go.

I also updated the Travis CI job to run tests on the latest versions of Go.

* Removing blank line for better import ordering

* Removing two accidentally added print statements