Increase flexibility of functions (#53)
In order to allow for greater flexibility in the `webauthn` package I
added the `CreateCredential` and `ValidateLogin` methods. Instead of
accepting a `http.Request` they now take a parsed response, so the
source can be up to the consumer of the package.

Aside from that I added two new methods in the `protocol` package:
`ParseCredentialRequestResponseBody` and
`ParseCredentialCreationResponseBody`. Instead of taking a
`http.Request` these methods accept a `io.Reader`.

The changes have been made with backwards compatibility in mind.