authenticator: use canonical cbor encoding to marshal credential public key
cbor.Marshal doesn't seem to be stable across runs.  It produces
different keyBytes on subsequent calls with the same input which is
extremely confusing if you're inspecting the attestation data directly
and comparing it to previous attestations.

Swap out for canonical Marshal as per Webauthn spec which uses a stable
ordering of map elements.