feat: discoverable credential login
This implements BeginDiscoverableLogin and ValidateDiscoverableLogin which can be used to easily perform discoverable credential logins (commonly known as passwordless login, and previously known as resident keys). This is implemented via an interface that allows ValidateDiscoverableLogin to lookup the user credentials either by the credential ID or userHandle as per the spec. See https://w3c.github.io/webauthn/#user-handle.